refacotr(osu-base): separate section handlers to their own parsers
