+++ packages/osu-base/src/constants/BeatmapSection.ts
@@ -1,0 +1,13 @@
+/**
+ * Beatmap sections that exist in an `.osu` file.
+ */
+export enum BeatmapSection {
+    general = "General",
+    editor = "Editor",
+    metadata = "Metadata",
+    difficulty = "Difficulty",
+    events = "Events",
+    timingPoints = "TimingPoints",
+    colors = "Colours",
+    hitObjects = "HitObjects",
+}
